Use passport.user.id instead of the whole object
I faced some weird issues in finding a user for an already registered passport. Turned out that the object that you pass to findOne is "normalized" by waterline. normalize.criteria(criteria) however is not (yet) able to cope with the associations correctly (at least that's what I guess). 

Passing in the id works perfectly fine for me. If you have a better suggestion or more details, please let me know.

However, AWESOME work with this generator! Thank you so much!!!!